edit server
서버 컨트롤러 추가

+++ APIGameServer/Repositories/Interfaces/IRedisDatabase.cs
@@ -3,6 +3,6 @@
 
 public interface IRedisDatabase : IDisposable
 {
-    public Task<long> GetNewUserUid();
+    public Task<int> GetNewUserUid();
 
 }
